requirements.txt
fastapi
requests
beautifulsoup4
PyExecJS
uvicorn
